#48 Rewrite invite api routes to used cached db

+++ src/pages/api/invite/[token].js
@@ -1,44 +1,38 @@
-import nc from 'next-connect';
 import jwt from 'next-auth/jwt';
-import middleware from '../../../middlewares/middleware';
+import { connectToDatabase } from '../../../utils/dbUtil';
 
 const secret = process.env.AUTH_SECRET;
 
-const handler = nc()
-  .use(middleware)
-  .get(
-    async (req, res) => {
+
+const handler = async (req, res) => {
+  const { method } = req;
+  if (method === 'GET') {
+    const { token } = req.query;
+    const { db } = await connectToDatabase();
+    const doc = await db
+      .collection('inviteTokens')
+      .find(
+        {
+          token,
+        },
+      ).toArray();
+    if (doc[0]) res.status(200).json(doc[0]);
+    else res.status(404).end('Not Found');
+  } else if (method === 'DELETE') {
+    const jwtTok = await jwt.getToken({ req, secret });
+    if (jwtTok && jwtTok.exp > 0) {
       const { token } = req.query;
-      await req.db
+      const { db } = await connectToDatabase();
+      const doc = await db
         .collection('inviteTokens')
-        .findOne(
+        .findOneAndDelete(
           {
             token,
           },
-          (err, doc) => {
-            if (err) throw err;
-            res.status(200).json(doc);
-          },
         );
-    },
-  ).delete(
-    async (req, res) => {
-      const jwtTok = await jwt.getToken({ req, secret });
-      if (jwtTok && jwtTok.exp > 0) {
-        const { token } = req.query;
-        await req.db
-          .collection('inviteTokens')
-          .findOneAndDelete(
-            {
-              token,
-            },
-            (err, doc) => {
-              if (err) throw err;
-              res.status(200).json(doc);
-            },
-          );
-      } else res.status(403).json({ error: '403 Invalid or expired token' });
-    },
-  );
+      res.status(200).json(doc);
+    } else res.status(403).end('403 Invalid or expired token');
+  } else res.status(405).end(`Method ${method} Not Allowed`);
+};
 
 export default handler;
